+++ packages/eas-cli/src/commands/account/login.ts
@@ -1,7 +1,10 @@
-import { Flags } from '@oclif/core';
+import { Errors, Flags } from '@oclif/core';
+import chalk from 'chalk';
 
 import EasCommand from '../../commandUtils/EasCommand';
 import Log from '../../log';
+import { confirmAsync } from '../../prompts';
+import { getActorDisplayName } from '../../user/User';
 
 export default class AccountLogin extends EasCommand {
   static override description = 'log in with your Expo account';
@@ -17,6 +20,7 @@ export default class AccountLogin extends EasCommand {
   };
 
   static override contextDefinition = {
+    ...this.ContextOptions.MaybeLoggedIn,
     ...this.ContextOptions.SessionManagment,
   };
 
@@ -25,7 +29,28 @@ export default class AccountLogin extends EasCommand {
       flags: { sso },
     } = await this.parse(AccountLogin);
 
-    const { sessionManager } = await this.getContextAsync(AccountLogin, { nonInteractive: false });
+    const {
+      sessionManager,
+      maybeLoggedIn: { actor },
+    } = await this.getContextAsync(AccountLogin, { nonInteractive: false });
+
+    if (sessionManager.getAccessToken()) {
+      throw new Error(
+        'EXPO_TOKEN is set in your environment, and is being used for all EAS authentication. To use username/password authentication, unset EXPO_TOKEN in your environment and re-run the command.'
+      );
+    }
+
+    if (actor) {
+      Log.warn(`You are already logged in as ${chalk.bold(getActorDisplayName(actor))}.`);
+
+      const shouldContinue = await confirmAsync({
+        message: 'Do you want to continue?',
+      });
+      if (!shouldContinue) {
+        Errors.error('Aborted', { exit: 1 });
+      }
+    }
+
     await sessionManager.showLoginPromptAsync({ sso });
     Log.log('Logged in');
   }