+++ packages/eas-cli/src/project/customBuildConfig.ts
@@ -1,6 +1,6 @@
 import { Platform } from '@expo/eas-build-job';
 import { BuildProfile } from '@expo/eas-json';
-import { errors, readAndValidateBuildConfigAsync } from '@expo/steps';
+import { errors, readAndValidateBuildConfigFromPathAsync } from '@expo/steps';
 import chalk from 'chalk';
 import fs from 'fs-extra';
 import path from 'path';
@@ -31,7 +31,9 @@ export async function validateCustomBuildConfigAsync({
       `Custom build configuration file ${chalk.bold(relativeConfigPath)} does not exist.`
     );
   }
-  if (await vcsClient.isFileIgnoredAsync(relativeConfigPath)) {
+
+  const rootDir = path.normalize(await vcsClient.getRootPathAsync());
+  if (await vcsClient.isFileIgnoredAsync(path.relative(rootDir, configPath))) {
     throw new Error(
       `Custom build configuration file ${chalk.bold(
         relativeConfigPath
@@ -40,7 +42,7 @@ export async function validateCustomBuildConfigAsync({
   }
 
   try {
-    const config = await readAndValidateBuildConfigAsync(configPath, {
+    const config = await readAndValidateBuildConfigFromPathAsync(configPath, {
       skipNamespacedFunctionsOrFunctionGroupsCheck: true,
     });
     return {